requirements.txt
django==3.2.13
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
psycopg2==2.9.2
pillow==9.0.1
djangorestframework-simplejwt==5.2.0
pyotp==2.6.0
dj_database_url
django_cors_headers
whitenoise
gunicorn
